feat: move move packet to entities namespace
