+++ Core/NosSmooth.Packets/Converters/Packets/InPacketConverter.cs
@@ -4,6 +4,7 @@
 //  Copyright (c) František Boháček. All rights reserved.
 //  Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using NosSmooth.Packets.Packets.Server.Entities;
 using NosSmooth.Packets.Packets.Server.Map;
 using Remora.Results;
